### 2020-08-25 #2567 @harpolea
@zpace Thanks for your response! We typically ask that reviews are completed within 6 weeks – do you think that you'd be able to complete a review within that time frame?
