### 2020-08-28 #2608 @whedon
[REVIEW]: AstroPaint: A Python Package for Painting Halo Catalogs into Celestial Maps
**Submitting author:** @syasini (<a href="http://orcid.org/0000-0003-1978-6325">Siavash Yasini</a>)
**Repository:** <a href="https://github.com/syasini/AstroPaint" target ="_blank">https://github.com/syasini/AstroPaint</a>
**Version:** 0.1.1
