### 2020-09-01 syasini/AstroPaint#70 @zpace
add LICENSE file
Hello, I am one of your JOSS reviewers! (https://github.com/openjournals/joss-reviews/issues/2608)

I haven't been able to locate a plain-text LICENSE file in this repo. I see that an MIT license is listed in setup.py, so it would suffice to add the text from [here](https://opensource.org/licenses/MIT) into a new file called LICENSE in the top-level directory.
