### 2020-09-01 syasini/AstroPaint#70 @syasini
Hi Zach. Thanks for pointing that out! I don't know how I forgot to add the file to the repo.

I just pushed a new commit to `develop` branch. Please check out the license file [here](https://github.com/syasini/AstroPaint/blob/develop/LICENSE) and if it's satisfactory I'll merge it to master and will close this issue.
