### 2020-09-02 syasini/AstroPaint#72 @zpace
painter.spray(..., parallel=True) fails ("unexpected keyword argument")
JOSS reviewer here, again! (https://github.com/openjournals/joss-reviews/issues/2608)

I'm working your tutorial notebook, and I have run into a problem with `Painter.spray()`'s `parallel=True` argument. It seems that `parallel=True` is being passed to the function used as the `template` argument of the `Painter` instance. This is happening because `Painter.spray()` has no argument by that name. Eliminating this argument from the call to `Painter.spray()` yields the expected output.

I'm on version 0.1.1, which appears to be the most current. FWIW, this error also manifests with the public binder.

I *suspect* that the `n_cpus` argument has replaced `parallel`. If this is correct, please do update all public tutorials to reflect this.
