### 2020-09-02 syasini/AstroPaint#73 @zpace
public documentation: Statement of need
JOSS reviewer here! (https://github.com/openjournals/joss-reviews/issues/2608)

Please add a Statement of Need to the public-facing documentation of this repo. You could paraphrase the text already included in the JOSS manuscript submission; and then insert it into either your existing repo README, include it in the RTD homepage (see https://github.com/syasini/AstroPaint/issues/54), or preferably both!
