### 2020-09-03 syasini/AstroPaint#72 @syasini
Thanks for catching this bug @zpace. You're absolutely correct. `n_cpus` was suggested as a replacement for the `parallel` argument to give the user more control over parallelization, but after implementing the change we did not update the tutorial notebook.

I just pushed a corrected version to `develop` which you can see here:   https://github.com/syasini/AstroPaint/blob/develop/tutorial.ipynb

(I really don't like GitHub's new notebook formatting; I really hope it's part of an A/B test and will be rolled out soon!)

I also fixed another error in a catalog file name at the very end.

I'll check the other notebooks for similar errors and will keep you updated. Meanwhile, let me know if this looks good so I can merge it to master.
