### 2020-09-03 syasini/AstroPaint#72 @zpace
I think there's also a case of this in the README. Feel free to close this after checking the tutorial and other example notebooks.
