### 2020-09-03 syasini/AstroPaint#72 @syasini
You're absolutely right! I'll fix that one as well. Will close after all the corrections are implemented.
