### 2020-09-03 c6b8160
Update parallel with n_cpus
