### 2020-09-03 9ac0817
Add new section: what is astropaint?
