+++ README.md
@@ -50,6 +50,17 @@ That's it! Now you can check out your masterpiece using
 
 ![BG](images/BG_websky_cover.png)
 
+# What is AstroPaint?
+
+AstroPaint is a python package for generating and visualizing sky maps of a wide range of astrophysical signals 
+originating from dark matter halos or the gas that they host. AstroPaint creates a whole-sky mock map of the 
+target signal/observable, at a desired resolution, by combining an input halo catalog and the radial/angular 
+profile of the astrophysical effect. The package also provides a suite of tools that can facilitate analysis
+ routines such as catalog filtering, map manipulation, and cutout stacking. The simulation suite has an 
+ Object-Oriented design and runs in parallel, making it both easy to use and readily scalable for production 
+ of high resolution maps with large underlying catalogs. Although the package has been primarily developed 
+ to simulate signals pertinent to galaxy clusters, its application extends to halos of arbitrary size or 
+ even point sources.
 
 # Package Structure 
 
