### 2020-09-03 syasini/AstroPaint#73 @syasini
For the moment I've added a new section `What is AstroPaint?` to the README file with the high-level overview of what the package is for. I will update this with a link to the JOSS paper after publication. I agree that this should definitely also go on the RTD homepage so I will keep that in mind.
