### 2020-09-03 syasini/AstroPaint#73 @syasini
I forgot to mention that the changes are applied to the `develop` branch as usual. I won't close this until I merge to `master`.
