### 2020-09-08 e4d9826
add CMB-HD noise configs.
