### 2020-09-08 a9f74d9
Fix bug in cutouts: Allow apply_func kwargs to have different lengths
