### 2020-09-14 4bbe781
Decrease image sizes
