+++ docs/index.rst
@@ -2,30 +2,343 @@
    sphinx-quickstart on Tue Sep  8 14:47:06 2020.
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
+.. role:: raw-html-m2r(raw)
+   :format: html
 
-AstroPaint's documentation
-==========================
 
-Introduction
-------------
+.. image:: docs/images/logo.PNG
+   :target: docs/images/logo.PNG
+   :height: 150
 
-Statement of Need
+AstroPaint
+==========
+
+*A python package for painting the sky*
+
+
+.. image:: https://mybinder.org/badge_logo.svg
+   :target: https://mybinder.org/v2/gh/syasini/AstroPaint/master?filepath=tutorial.ipynb
+   :alt: Binder
+
+
+You can install **AstroPaint** by running the following in the command line:
+
+``git clone https://github.com/syasini/AstroPaint.git``
+
+``cd AstroPaint``
+
+``pip install [-e] .``
+
+the ``-e`` argument will install the package in editable mode which is suitable for developement. If you want to modify the code use this option.
+
+**Important Note**:
+If you want the sample catalogs to be cloned automatically along with the
+rest of the repository, make sure you have `Git Large File Storage (git lfs) <https://git-lfs.github.com/>`_ installed.
+
+If you are a conda user, please consider creating a new environment before installation:
+
+ ``conda create -n astropaint python=3.7``
+
+ ``conda activate astropaint``
+
+Workflow
+========
+
+Converting catalogs to mock maps with AstroPaint is extremely simple. Here is what an example session looks like:
+
+.. code-block:: python
+
+   from astropaint import Catalog, Canvas, Painter
+
+   catalog = Catalog(data=your_input_data)
+
+   canvas = Canvas(catalog, nside)
+
+   painter = Painter(template=your_radial_profile)
+
+   painter.spray(canvas)
+
+That's it! Now you can check out your masterpiece using
+
+``canvas.show_map()``
+
+
+.. image:: docs/images/BG_websky_cover.png
+   :target: docs/images/BG_websky_cover.png
+   :alt: BG
+
+What is AstroPaint?
+===================
+AstroPaint is a python package for generating and visualizing sky maps of a wide range of astrophysical signals
+originating from dark matter halos or the gas that they host. AstroPaint creates a whole-sky mock map of the target
+signal/observable, at a desired resolution, by combining an input halo catalog and the radial/angular profile of the
+astrophysical effect. The package also provides a suite of tools that can facilitate analysis routines such as catalog
+filtering, map manipulation, and cutout stacking. The simulation suite has an Object-Oriented design and runs in
+parallel, making it both easy to use and readily scalable for production of high resolution maps with large underlying
+catalogs. Although the package has been primarily developed to simulate signals pertinent to galaxy clusters, its
+application extends to halos of arbitrary size or even point sources.
+
+Package Structure
+=================
+
+While there is no external documentation for the code yet, you can use `this chart <https://www.mindmeister.com/1417665103/astropaint-astropaint-py?fullscreen=1>`_
+to understand the package structure and see what methods are available so far.
+
+Examples
+========
+
+Nonsense Template
 -----------------
 
-Features
+Here's an example script that paints a nonsense template on a 10 x 10 [sqr deg]
+patch of the ``Sehgal`` catalog:
+
+.. code-block:: python
+
+   import numpy as np
+   from astropaint import Catalog, Canvas, Painter
+
[…]
