+++ docs/conf.py
@@ -29,6 +29,8 @@ author = 'Siavash Yasini'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
+
+master_doc = 'index'
 extensions = ['recommonmark', 'sphinx.ext.autodoc', 'sphinx.ext.napoleon']
 
 napoleon_google_docstring = False
