+++ docs/index.rst
@@ -6,8 +6,8 @@
    :format: html
 
 
-.. image:: docs/images/logo.PNG
-   :target: docs/images/logo.PNG
+.. image:: ./images/logo.PNG
+   :target: ./images/logo.PNG
    :height: 150
 
 AstroPaint
@@ -63,8 +63,8 @@ That's it! Now you can check out your masterpiece using
 ``canvas.show_map()``
 
 
-.. image:: docs/images/BG_websky_cover.png
-   :target: docs/images/BG_websky_cover.png
+.. image:: ./images/BG_websky_cover.png
+   :target: ./images/BG_websky_cover.png
    :alt: BG
 
 What is AstroPaint?
@@ -121,8 +121,8 @@ patch of the ``Sehgal`` catalog:
 
 
 
-.. image:: docs/images/a_random_template.png
-   :target: docs/images/a_random_template.png
+.. image:: ./images/a_random_template.png
+   :target: ./images/a_random_template.png
    :height: 300
    :align: center
 
@@ -139,8 +139,8 @@ with. Let's spray ths canvas now:
    canvas.show_map("cartview", lonra=[0,10], latra=[0,10])
 
 
-.. image:: docs/images/a_random_map.png
-   :target: docs/images/a_random_map.png
+.. image:: ./images/a_random_map.png
+   :target: ./images/a_random_map.png
    :height: 300
    :align: center
 
@@ -152,8 +152,8 @@ The default value `n_cpus=-1` uses all the available cpus.
 
 
 
-.. image:: docs/images/parallel.gif
-   :target: docs/images/parallel.gif
+.. image:: ./images/parallel.gif
+   :target: ./images/parallel.gif
    :height: 450
    :align: center
 
@@ -176,8 +176,8 @@ You can easily stack cutouts of the map using the following:
 
 
 
-.. image:: docs/images/a_random_stack.png
-   :target: docs/images/a_random_stack.png
+.. image:: ./images/a_random_stack.png
+   :target: ./images/a_random_stack.png
    :height: 200
    :align: center
 
@@ -246,7 +246,7 @@ This is much faster, but the speed comes at a small price. If your 3D profile
 is not smooth, the interpolated 2D projection will slightly deviate from the
 exact integration.
 
-.. image:: docs/images/tophat_interp.png
+.. image:: ./images/tophat_interp.png
    :align: center
    :height: 200
 
@@ -270,8 +270,8 @@ kinetic Sunyaev-Zeldovich (kSZ) on the WebSky catalog halos.
 
 Since the shape of the profile is smooth, we won't lose accuracy by using the interpolator.
 
-.. image:: docs/images/battaglia16_tau.png
-   :target: docs/images/battaglia16_tau.png
+.. image:: ./images/battaglia16_tau.png
+   :target: ./images/battaglia16_tau.png
    :height: 200
    :align: center
 
@@ -290,8 +290,8 @@ cut of 8E13 M_sun.
    tau_painter.spray(canvas)
 
 
-.. image:: docs/images/tau_map_battaglia.png
-   :target: docs/images/tau_map_battaglia.png
+.. image:: ./images/tau_map_battaglia.png
+   :target: ./images/tau_map_battaglia.png
    :height: 200
    :align: center
 
@@ -305,8 +305,8 @@ dimensionless velocity of the halos to get the kSZ signal.
 
 And here is what it looks like:
 
-.. image:: docs/images/ksz_map_battaglia.png
-   :target: docs/images/ksz_map_battaglia.png
+.. image:: ./images/ksz_map_battaglia.png
+   :target: ./images/ksz_map_battaglia.png
    :height: 200
    :align: center
 
[…]
