### 2020-09-14 39e4b48
Increase image sizes
