### 2020-09-14 bafd130
Fix art gallery image positions
