### 2020-09-16 817d613
change logo size
