### 2020-09-16 c59f3de
Add documentation link to README and index.rst
