+++ README.md
@@ -64,7 +64,7 @@ profile of the astrophysical effect. The package also provides a suite of tools 
 
 # Package Structure 
 
-While there is no external documentation for the code yet, you can use [this
+See our [documentation](https://astropaint.readthedocs.io/) and [this
  chart](https://www.mindmeister.com/1417665103/astropaint-astropaint-py?fullscreen=1)
  to understand the package structure and see what methods are available so
   far. 
