### 2020-09-16 syasini/AstroPaint#73 @syasini
Closing this (see #54 ).
