### 2020-09-18 syasini/AstroPaint#72 @syasini
both instances of the `parallel=True` in the tutorial and README have been fixed. Thanks again for catching the mistakes!

Closing this.
