### 2020-09-21 dda70eb
Fix flake8 errors: change is to ==
