+++ astropaint/paint_bucket.py
@@ -1174,7 +1174,7 @@ class Canvas:
 
         def gen_center_ipix(self, halo_list="All"):
             """generate ipix of the halo centers"""
-            if halo_list is "All":
+            if halo_list =="All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1188,7 +1188,7 @@ class Canvas:
             """generate the angle (theta, phi) of the halo centers
             if snap2pixel is True, the angular coordinate of the halo center pixel will be
             returned"""
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1203,7 +1203,7 @@ class Canvas:
                            self.catalog.data.phi[halo])
 
         def gen_center_vec(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1213,7 +1213,7 @@ class Canvas:
                                  self.catalog.data.phi[halo])
 
         def gen_pixel_index(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1229,7 +1229,7 @@ class Canvas:
                               )
 
         def gen_pixel_ang(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1245,7 +1245,7 @@ class Canvas:
             -------
             None
             """
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1254,7 +1254,7 @@ class Canvas:
                 yield np.asarray(hp.pix2vec(self.nside, index)).T
 
         def gen_cent2pix_rad(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1264,7 +1264,7 @@ class Canvas:
                 yield hp.rotator.angdist(np.squeeze(pixel_ang), center_ang)
 
         def gen_cent2pix_mpc(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             assert hasattr(halo_list, '__iter__')
@@ -1273,7 +1273,7 @@ class Canvas:
                 yield self.center_D_a[halo] * pix2cent_rad
 
         def gen_cent2pix_hat(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             for (pix_vec, cent_vec) in zip(self.gen_pixel_vec(halo_list),
@@ -1281,7 +1281,7 @@ class Canvas:
                 yield Canvas._normalize_vec(pix_vec - cent_vec)
 
         def gen_cent2pix_mpc_vec(self, halo_list="All"):
-            if halo_list is "All":
+            if halo_list == "All":
                 halo_list = range(self.catalog.size)
 
             for (halo, pix_vec, cent_vec) in zip(halo_list,
@@ -1446,7 +1446,8 @@ class Canvas:
 
         [set_to_1(disc) for disc in self.discs.gen_pixel_index()]
 
[…]
