### 2020-09-21 1cde654
Fix halo_list == "all" bug: comparing list to string.
