+++ .github/workflows/python-package.yml
@@ -7,7 +7,7 @@ on:
   push:
     branches: [ develop, master ]
   pull_request:
-    branches: [ master ]
+    branches: [ develop, master ]
 
 jobs:
   build:
