### 2020-09-22 #2608 @zpace
After having reviewed this submission, I am happy to recommend it for acceptance into JOSS. The developers have responded to my handful of comments and suggestions to my satisfaction.
