### 2020-10-15 4908b9c
Change the default value of n_cpus to -1"
