### 2020-10-15 adbc78e
change default n_cpus=-1 in README
