+++ README.rst
@@ -1,19 +1,15 @@
-.. AstroPaint documentation master file, created by
-   sphinx-quickstart on Tue Sep  8 14:47:06 2020.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
 .. role:: raw-html-m2r(raw)
    :format: html
 
 
-.. image:: ./images/logo.PNG
-   :target: ./images/logo.PNG
-   :height: 250
+.. image:: docs/images/logo.PNG
+   :target: docs/images/logo.PNG
+   :height: 100
 
 AstroPaint
 ==========
 
-*A python package for painting the sky*
+*A python package for painting the sky* 
 
 
 .. image:: https://mybinder.org/badge_logo.svg
@@ -27,7 +23,7 @@ You can install **AstroPaint** by running the following in the command line:
 
 ``cd AstroPaint``
 
-``pip install [-e] .``
+``pip install [-e] .`` 
 
 the ``-e`` argument will install the package in editable mode which is suitable for developement. If you want to modify the code use this option.
 
@@ -63,8 +59,8 @@ That's it! Now you can check out your masterpiece using
 ``canvas.show_map()``
 
 
-.. image:: ./images/BG_websky_cover.png
-   :target: ./images/BG_websky_cover.png
+.. image:: docs/images/BG_websky_cover.png
+   :target: docs/images/BG_websky_cover.png
    :alt: BG
 
 What is AstroPaint?
@@ -81,7 +77,7 @@ application extends to halos of arbitrary size or even point sources.
 Package Structure
 =================
 
-See our `documentation <https://astropaint.readthedocs.io/>`_ and `this chart <https://www.mindmeister.com/1417665103/astropaint-astropaint-py?fullscreen=1>`_
+While there is no external documentation for the code yet, you can use `this chart <https://www.mindmeister.com/1417665103/astropaint-astropaint-py?fullscreen=1>`_
 to understand the package structure and see what methods are available so far.
 
 Examples
@@ -115,14 +111,14 @@ patch of the ``Sehgal`` catalog:
    # pass the template to the painter
    painter = Painter(template=a_nonsense_template)
 
-   # plot the template for halos #0, #10, and #100 for R between 0 to 5 Mpc
+   # plot the template for halos #0, #10, and #100 for R between 0 to 5 Mpc 
    R = np.linspace(0,5,100)
    painter.plot_template(R, catalog, halo_list=[0,10,100])
 
 
 
-.. image:: ./images/a_random_template.png
-   :target: ./images/a_random_template.png
+.. image:: docs/images/a_random_template.png
+   :target: docs/images/a_random_template.png
    :height: 300
    :align: center
 
@@ -139,8 +135,8 @@ with. Let's spray ths canvas now:
    canvas.show_map("cartview", lonra=[0,10], latra=[0,10])
 
 
-.. image:: ./images/a_random_map.png
-   :target: ./images/a_random_map.png
+.. image:: docs/images/a_random_map.png
+   :target: docs/images/a_random_map.png
    :height: 300
    :align: center
 
@@ -152,8 +148,8 @@ Setting `n_cpus=-1` uses all the available cpus.
 
 
 
-.. image:: ./images/parallel.gif
-   :target: ./images/parallel.gif
+.. image:: docs/images/parallel.gif
+   :target: docs/images/parallel.gif
    :height: 450
    :align: center
 
@@ -176,9 +172,9 @@ You can easily stack cutouts of the map using the following:
 
 
 
-.. image:: ./images/a_random_stack.png
-   :target: ./images/a_random_stack.png
-   :height: 300
+.. image:: docs/images/a_random_stack.png
[…]
