### 2020-10-15 904f300
Remove README.rst
