### 2020-10-15 a7e96bc
Change websky_lite_redshift to WebSky_lite
