+++ docs/index.rst
@@ -281,7 +281,7 @@ cut of 8E13 M_sun.
 
 .. code-block:: python
 
-   catalog = Catalog("websky_lite_redshift")
+   catalog = Catalog("WebSky_lite")
    catalog.cut_lon_lat(lon_range=[5,10], lat_range=[5,10])
    catalog.cut_M_200c(8E13)
 
