### 2020-10-19 f16e083
relax the version requirements in the dependencies from == to >=
