+++ requirements.txt
@@ -1,16 +1,16 @@
-astropy==4.0.0
-matplotlib==3.1.2
-numpy==1.19.1
-pytest==5.3.1
-scipy==1.3.0
-pandas==1.0.1
-tqdm==4.48.0
-pyyaml==5.1.2
-decorator==4.4.1
-seaborn==0.9.0
-joblib==0.14.0
-requests==2.24.0
+astropy>=4.0.0
+matplotlib>=3.1.2
+numpy>=1.19.1
+pytest>=5.3.1
+scipy>=1.3.0
+pandas>=1.0.1
+tqdm>=4.48.0
+pyyaml>=5.1.2
+decorator>=4.4.1
+seaborn>=0.9.0
+joblib>=0.14.0
+requests>=2.24.0
 jupyter
-requests==2.24.0
+requests>=2.24.0
 healpy==1.13.0
 
