### 2020-10-20 #2608 @AshKelly
I am happy to recommend this submission for acceptance into JOSS. The developers have responded positively to all of my comments and suggestions.

I would also like to add that I really like the library and the simplicity of the API. Notably, the abstraction into three main components `Catalog`, `Canvas` and `Painter`. I found that after just a short period experimenting with examples given in the docs, I was able to produce quite involved catalogues with relative ease. I anticipate this is a library I will use in the future. Great job!

I'd also like to thank @zpace as I benefited from the documentation that they helped produce. The RTD is a really great addition!
