### 2020-10-30 #2608 @arfon
@syasini - I made a tiny formatting fix in https://github.com/syasini/AstroPaint/pull/85. After merging (assuming you like the change), could you make a new release of this software that includes the changes that have resulted from this review. Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
