### 2020-11-02 #2608 @syasini
Thanks for your help @arfon. I just commented on issue #85. I sincerely appreciate your suggestion.

After you submit the PR to `develop` I will go ahead with the new version release and submission to Zenodo and will report back here with the updates.
