### 2020-11-04 #2608 @syasini
@arfon, I published a new release on (v1.0.0) on Zenodo. You can find it [here](https://zenodo.org/record/4243176#.X6I9qpNKjt1). Let me know if further actions are needed on my side.

And thanks for resubmitting the PR to `develop`. I merged it earlier today :)
