+++ paper/paper.bib
@@ -28,7 +28,7 @@ archivePrefix = {arXiv},
         month = jan,
           eid = {arXiv:2001.08787},
         pages = {arXiv:2001.08787},
-          doi = {10.1088/1475-7516/2020/10/012}
+          doi = {10.1088/1475-7516/2020/10/012},
 archivePrefix = {arXiv},
        eprint = {2001.08787},
  primaryClass = {astro-ph.CO},
