### 2020-11-07 #2608 @syasini
Thanks for the great news @arfon! This is very exciting.

And many thanks again to @zpace and @AshKelly for their insightful feedback and their positive reviews and to @harpolea for editing.

Publishing with JOSS was a great experience and we'll be looking forward to hopefully do it again in the future!

Thank you all 🙏
